+++ Connections/MSPBSL_Connection_v1_10.h
@@ -61,8 +61,8 @@ public:
 	virtual uint16_t RX_Password(void);
 	virtual uint16_t massErase(void);
 	virtual uint16_t TX_BSL_Version(string& versionString);
-	virtual uint16_t eraseCheck( uint16_t startAddr, uint32_t numBytes );
-	//virtual uint16_t InfoMainErase(uint16_t addr);  //Workaround possible, but not able to reproduce bug
+	virtual uint16_t eraseCheck( uint32_t startAddr, uint32_t numBytes );
+	//virtual uint16_t InfoMainErase(uint32_t addr);  //Workaround possible, but not able to reproduce bug
 
 	virtual string getErrorInformation( uint16_t err );
